rename components directory to work on github error of extra components (Components) directory
